requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow==9.3.0
django-cors-headers